Rewrote seatbelt
The Buckled() function creates a thread that runs in the background. This thread continuously checks the value of the variable isBuckled. As long as isBuckled is true, the thread will use the function DisableControlAction(0, 75, true) to disable the player's ability to exit from the vehicle.

When the player toggles their seatbelt on or off, the Seatbelt(status) function is called.
If the player is buckling up, the function sends a message to the NUI to update the seatbelt indicator on the screen, uses the native function to set the windscreen parameters so the player doesn't fall out on the windshield, and calls the Buckled() function.
If the player unbuckles, the function sends a message to the NUI to update the seatbelt indicator on the screen, uses the native function to set the windscreen parameters so the player can fall out of the windscreen, and sets the isBuckled variable to false. 

The script also creates a background thread that runs indefinitely.
When the player enters or exits a vehicle, this thread checks the player's current vehicle and updates the NUI's seatbelt indicator to match the value of isBuckled.

In addition, the script registers two game commands:


The command dolu hud:seatbelt toggles the seatbelt on and off.

The RegisterKeyMapping('dolu hud:seatbelt', 'Toggle seatbelt', 'keyboard', Config.seatbeltKey) command associates the seatbelt toggle command with a specific key on the player's keyboard.


When the player presses the key, the script determines whether the player is in a vehicle, and if the current vehicle class is not a motorcycle, cycle or boat, the script invokes the Seatbelt(not isBuckled) function, which toggles the seatbelt on or off and updates the NUI and windscreen parameters accordingly.